by zx

+++ topology/fc_topo_gene_cpp/src/fc_tagger_test.h
@@ -19,8 +19,8 @@ class FcTaggerTest: public Fc_base{
         // Save the topo infor in files for ksp function interface 
         void SaveTaggerGraph();
 
-        uint16_t SearchKsp(int srcIn, int dstIn, int pathNum, int thLabel, uint16_t *pathInfor);
-        uint16_t SearchEcmp(int srcIn, int dstIn, int thLabel, uint16_t *pathInfor);
+        uint16_t SearchKsp(int srcIn, int dstIn, int pathNum, int thLabel, int *pathInfor);
+        uint16_t SearchEcmp(int srcIn, int dstIn, int thLabel, int *pathInfor);
 
 
     private:
